+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/element/KeywordStatement.java
@@ -28,4 +28,6 @@ public interface KeywordStatement extends RobotStatement, PsiNamedElement, StubB
 
    @Nullable
    DefinedVariable getGlobalVariable();
+
+   void reset();
 }